backend/requirements.txt
fastapi>=0.124.0,<0.130.0
uvicorn[standard]>=0.38.0
pydantic>=2.7.0,<3.0.0
python-dotenv>=1.2.0
requests>=2.32.0
beautifulsoup4>=4.14.0
google-search-results>=2.4.0   # serpapi client (package name on PyPI)
